Moving retry logic/code into apiclient
