+++ cristin-organization/src/main/java/no/unit/nva/cristin/organization/CristinOrganizationApiClient.java
@@ -14,8 +14,6 @@ import nva.commons.apigateway.exceptions.NotFoundException;
 import nva.commons.core.attempt.Failure;
 import nva.commons.core.attempt.Try;
 import nva.commons.core.paths.UriWrapper;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.net.URI;
 import java.net.http.HttpClient;
@@ -54,13 +52,8 @@ public class CristinOrganizationApiClient extends ApiClient {
 
     public static final String CRISTIN_PER_PAGE_PARAM = "per_page";
     private static final String CRISTIN_QUERY_NAME_PARAM = "name";
-public static final int FIRST_EFFORT = 0;
-    public static final int MAX_EFFORTS = 2;
-    public static final int WAITING_TIME = 500; //500 milliseconds
-    public static final String LOG_INTERRUPTION = "InterruptedException while waiting to resend HTTP request";
     public static final String ERROR_MESSAGE_FORMAT = "%d:%s";
     public static final String NULL_HTTP_RESPONSE_ERROR_MESSAGE = "No HttpResponse found";
-    private static final Logger logger = LoggerFactory.getLogger(CristinOrganizationApiClient.class);
 
     public CristinOrganizationApiClient() {
         this(HttpClient.newBuilder()
@@ -72,7 +65,6 @@ public static final int FIRST_EFFORT = 0;
     public CristinOrganizationApiClient(HttpClient client) {
         super(client);
     }
-
 
     /**
      * Get information for an Organization.
@@ -254,45 +246,6 @@ public static final int FIRST_EFFORT = 0;
         return new FailedHttpRequestException(failureException, failureException.getStatusCode());
     }
 
-    protected Try<HttpResponse<String>> sendRequestMultipleTimes(URI uri) {
-        Try<HttpResponse<String>> lastEffort = null;
-        for (int effortCount = FIRST_EFFORT; shouldKeepTrying(effortCount, lastEffort); effortCount++) {
-            waitBeforeRetrying(effortCount);
-            lastEffort = attemptFetch(uri, effortCount);
-        }
-        return lastEffort;
-    }
-
-    private Try<HttpResponse<String>> attemptFetch(URI uri, int effortCount) {
-        Try<HttpResponse<String>> newEffort = attempt(() -> fetchGetResultAsync(uri).get());
-        if (newEffort.isFailure()) {
-            logger.warn(String.format("Failed HttpRequest on attempt %d of 3: ", effortCount + 1)
-                    + newEffort.getException().getMessage(), newEffort.getException()
-            );
-        }
-        return newEffort;
-    }
-
-    private boolean shouldTryMoreTimes(int effortCount) {
-        return effortCount < MAX_EFFORTS;
-    }
-
-    @SuppressWarnings("PMD.UselessParentheses") // keep the parenthesis for clarity
-    private boolean shouldKeepTrying(int effortCount, Try<HttpResponse<String>> lastEffort) {
-        return lastEffort == null || (lastEffort.isFailure() && shouldTryMoreTimes(effortCount));
-    }
-
-    private int waitBeforeRetrying(int effortCount) {
-        if (effortCount > FIRST_EFFORT) {
-            try {
-                Thread.sleep(WAITING_TIME);
-            } catch (InterruptedException e) {
-                logger.error(LOG_INTERRUPTION);
-                throw new RuntimeException(e);
-            }
-        }
-        return effortCount;
-    }
 
     protected SubSubUnitDto getSubSubUnitDtoWithMultipleEfforts(URI subunitUri)
             throws ApiGatewayException {
